cleanup(trace-eap-waterfall): Removing old quick trace queries
