+++ static/app/views/performance/transactionSummary/transactionTags/tagsHeatMap.tsx
@@ -7,7 +7,7 @@ import {useOverlay} from '@react-aria/overlays';
 import {useOverlayTriggerState} from '@react-stately/overlays';
 import {truncate} from '@sentry/core';
 import type {VisualMapComponentOption} from 'echarts';
-import type {Location} from 'history';
+import type {Location, LocationDescriptor} from 'history';
 import memoize from 'lodash/memoize';
 
 import HeatMapChart from 'sentry/components/charts/heatMapChart';
@@ -15,12 +15,12 @@ import {HeaderTitleLegend} from 'sentry/components/charts/styles';
 import TransitionChart from 'sentry/components/charts/transitionChart';
 import TransparentLoadingMask from 'sentry/components/charts/transparentLoadingMask';
 import LoadingIndicator from 'sentry/components/loadingIndicator';
+import MenuItem from 'sentry/components/menuItem';
 import {Overlay, PositionWrapper} from 'sentry/components/overlay';
 import Panel from 'sentry/components/panels/panel';
 import PerformanceDuration from 'sentry/components/performanceDuration';
 import Placeholder from 'sentry/components/placeholder';
 import QuestionTooltip from 'sentry/components/questionTooltip';
-import {DropdownItem, SectionSubtext} from 'sentry/components/quickTrace/styles';
 import Truncate from 'sentry/components/truncate';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
@@ -465,5 +465,52 @@ const StyledPanel = styled(Panel)`
 `;
 
 const StyledHeaderTitleLegend = styled(HeaderTitleLegend)``;
+
+const SectionSubtext = styled('div')`
+  color: ${p => p.theme.subText};
+  font-size: ${p => p.theme.fontSize.md};
+`;
+
+const StyledMenuItem = styled(MenuItem)<{width: 'small' | 'large'}>`
+  width: ${p => (p.width === 'large' ? '350px' : '200px')};
+
+  &:not(:last-child) {
+    border-bottom: 1px solid ${p => p.theme.innerBorder};
+  }
+`;
+
+const MenuItemContent = styled('div')`
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+`;
+
+type DropdownItemProps = {
+  children: React.ReactNode;
+  allowDefaultEvent?: boolean;
+  onSelect?: (eventKey: any) => void;
+  to?: LocationDescriptor;
+  width?: 'small' | 'large';
+};
+
+export function DropdownItem({
+  children,
+  onSelect,
+  allowDefaultEvent,
+  to,
+  width = 'large',
+}: DropdownItemProps) {
+  return (
+    <StyledMenuItem
+      data-test-id="dropdown-item"
+      to={to}
+      onSelect={onSelect}
+      width={width}
+      allowDefaultEvent={allowDefaultEvent}
+    >
+      <MenuItemContent>{children}</MenuItemContent>
+    </StyledMenuItem>
+  );
+}
 
 export default TagsHeatMap;
